Split card classes into separate modules
